requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
